client:ensure readdir_cache does not retain a dangling pointer to dn
When calling unlink to clear a Dentry, the address of the dentry stored
in the readdir_cache is not synchronously cleared. This creates a dangling
pointer, which may lead to unknown errors if the address is reused.

Synchronously clear the dentry pointer stored in the readdir_cache
when deleting a dentry from the dentries.

Signed-off-by: luckgaojq <gaojq0426@foxmail.com>